consolidated database query

+++ admin/datamodel/article.php
@@ -15,7 +15,7 @@
 
 	You should have received a copy of the GNU General Public License
 	along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
+*/
 
 require_once(STARTPATH.DBPATH.'db.php');
 require_once(STARTPATH.DATAMODELPATH.'number.php');
@@ -45,19 +45,19 @@ class Article {
     private $updated;
     private $filter;
 
-    const INSERT_SQL = 'insert into articles (id, number_id, category_id, indexnumber, published, title, subtitle, summary, body, commentsallowed, tag, metadescription, metakeyword, created, updated) values (#, #, #, #, #, ?, ?, ?, ?, #, ?, ?, ?, now(), now())';
-    const UPDATE_SQL = 'update articles set number_id = #, category_id = #, indexnumber = #, published = #, commentsallowed = #, title = ?, subtitle = ?, summary = ?, body = ?, tag = ?, metadescription = ?, metakeyword = ?, updated=now() where id = #';
-    const DELETE_SQL = 'delete from articles where id = # ';
-    const DELETE_USER_ARTICLE = 'delete from users_articles where article_id = # ';
-    const SELECT_BY_ID = 'select * from articles where id = #';
-    const SELECT_LAST_N = 'select * from articles ORDER BY updated DESC LIMIT # ';
-    const SELECT_BY_TITLE = 'select * from articles where title like ?';
-    const FIND_IN_ALL_TEXT_FIELDS = 'select * from articles where title like ? OR subtitle like ? OR summary like ? OR body like ? ';
-    const FIND_IN_ALL_TEXT_FIELDS_PUBLISHED_ARTICLES = 'select * from articles where (title like ? OR subtitle like ? OR summary like ? OR body like ?) AND published=1 ';
-    const SELECT_COMMENTS_PUB = 'select * from comments where published=1 AND article_id = # order by created DESC';
-    const SELECT_COMMENTS = 'select * from comments where article_id = # order by created DESC';
-    const SELECT_CATEGORY = 'select * from categories where id = # Limit 1';
-    const SELECT_NUMBER = 'select * from numbers where id = #';
+    const INSERT_SQL = 'insert into articles (id, number_id, category_id, indexnumber, published, title, subtitle, summary, body, commentsallowed, tag, metadescription, metakeyword, created, updated) values (@#@, @#@, @#@, @#@, @#@, @?@, @?@, @?@, @?@, @#@, @?@, @?@, @?@, now(), now())';
+    const UPDATE_SQL = 'update articles set number_id = @#@, category_id = @#@, indexnumber = @#@, published = @#@, commentsallowed = @#@, title = @?@, subtitle = @?@, summary = @?@, body = @?@, tag = @?@, metadescription = @?@, metakeyword = @?@, updated=now() where id = @#@';
+    const DELETE_SQL = 'delete from articles where id = @#@ ';
+    const DELETE_USER_ARTICLE = 'delete from users_articles where article_id = @#@ ';
+    const SELECT_BY_ID = 'select * from articles where id = @#@';
+    const SELECT_LAST_N = 'select * from articles ORDER BY updated DESC LIMIT @#@ ';
+    const SELECT_BY_TITLE = 'select * from articles where title like @?@';
+    const FIND_IN_ALL_TEXT_FIELDS = 'select * from articles where title like @?@ OR subtitle like @?@ OR summary like @?@ OR body like @?@ ';
+    const FIND_IN_ALL_TEXT_FIELDS_PUBLISHED_ARTICLES = 'select * from articles where (title like @?@ OR subtitle like @?@ OR summary like @?@ OR body like @?@) AND published=1 ';
+    const SELECT_COMMENTS_PUB = 'select * from comments where published=1 AND article_id = @#@ order by created DESC';
+    const SELECT_COMMENTS = 'select * from comments where article_id = @#@ order by created DESC';
+    const SELECT_CATEGORY = 'select * from categories where id = @#@ Limit 1';
+    const SELECT_NUMBER = 'select * from numbers where id = @#@';
     const SELECT_LAST = 'select * from articles order by indexnumber DESC Limit 1';
     const SELECT_ALL_PUB = 'select * from articles where published=1 order by indexnumber DESC';
     const SELECT_ALL = 'select * from articles order by id DESC';
@@ -65,11 +65,11 @@ class Article {
     const SELECT_BY_INDEXNUMBER = 'select indexnumber from articles order by indexnumber DESC';
     const SELECT_MAX_INDEXNUMBER = 'select max(indexnumber) from articles ';
     const SELECT_BY_ID_ORD = 'select id from articles order by id DESC';
-    const SELECT_UP_INDEXNUMBER = 'select * from articles WHERE number_id = # AND indexnumber > # order by indexnumber ';
-    const SELECT_DOWN_INDEXNUMBER = 'select * from articles WHERE number_id = # AND indexnumber < # order by indexnumber DESC ';
-    const SELECT_USERS = 'select US.* from users as US, users_articles as UA where US.id = UA.user_id AND UA.article_id = # order by US.name ';
-    const DELETE_USER = 'delete from users_articles WHERE article_id = # AND user_id = # ';
-    const LINK_USER = 'insert into users_articles (article_id, user_id) values (#, #) ';
+    const SELECT_UP_INDEXNUMBER = 'select * from articles WHERE number_id = @#@ AND indexnumber > @#@ order by indexnumber ';
+    const SELECT_DOWN_INDEXNUMBER = 'select * from articles WHERE number_id = @#@ AND indexnumber < @#@ order by indexnumber DESC ';
+    const SELECT_USERS = 'select US.* from users as US, users_articles as UA where US.id = UA.user_id AND UA.article_id = @#@ order by US.name ';
+    const DELETE_USER = 'delete from users_articles WHERE article_id = @#@ AND user_id = @#@ ';
+    const LINK_USER = 'insert into users_articles (article_id, user_id) values (@#@, @#@) ';
 
 
     public function __construct($id=self::NEW_ARTICLE, $number_id=self::NEW_ARTICLE, $category_id=self::NEW_ARTICLE, $indexnumber='', $published='0', $title='', $subtitle='', $summary='', $body='', $commentsallowed='0', $tag='', $metadescription='', $metakeyword='', $created='', $updated='') {
@@ -99,10 +99,10 @@ class Article {
         try {
             $tables = array("articles" => TBPREFIX."articles");
             $rs = DB::getInstance()->execute(
-                $SQL,
-                $array_str,
-                $array_int,
-                $tables);
+                    $SQL,
+                    $array_str,
+                    $array_int,
+                    $tables);
             if ($row = mysql_fetch_array($rs)) {
                 $ret = new Article($row['id'], $row['number_id'], $row['category_id'], $row['indexnumber'], $row['published'], $row['title'], $row['subtitle'], $row['summary'], $row['body'], $row['commentsallowed'], $row['tag'], $row['metadescription'], $row['metakeyword'], $row['created'], $row['updated']);
             } else {
@@ -120,10 +120,10 @@ class Article {
         try {
             $tables = array("articles" => TBPREFIX."articles");
             $rs = DB::getInstance()->execute(
-                $SQL,
-                $array_str,
-                $array_int,
-                $tables);
+                    $SQL,
+                    $array_str,
+                    $array_int,
+                    $tables);
             $ret = array();
             while ($row = mysql_fetch_array($rs)) {
                 $ret[] = new Article($row['id'], $row['number_id'], $row['category_id'], $row['indexnumber'], $row['published'], $row['title'], $row['subtitle'], $row['summary'], $row['body'], $row['commentsallowed'], $row['tag'], $row['metadescription'], $row['metakeyword'], $row['created'], $row['updated']);
@@ -210,26 +210,26 @@ class Article {
         $ret = array();
         try {
             $tables = array('users' => TBPREFIX.'users',
-                'users_articles' => TBPREFIX.'users_articles');
+                    'users_articles' => TBPREFIX.'users_articles');
             $rs = DB::getInstance()->execute(
-                self::SELECT_USERS,
-                array(),
-                array($this->id),
-                $tables);
+                    self::SELECT_USERS,
+                    array(),
+                    array($this->id),
+                    $tables);
             while ($row = mysql_fetch_array($rs)) {
                 $ret[] = new User(
-                    $row['id'],
-                    $row['name'],
-                    $row['username'],
-                    $row['password'],
-                    $row['body'],
-                    $row['role'],
-                    $row['toshow'],
-                    $row['email'],
-                    $row['msn'],
-                    $row['skype'],
-                    $row['created'],
-                    $row['updated']);
+                        $row['id'],
+                        $row['name'],
+                        $row['username'],
+                        $row['password'],
+                        $row['body'],
+                        $row['role'],
+                        $row['toshow'],
+                        $row['email'],
+                        $row['msn'],
+                        $row['skype'],
+                        $row['created'],
+                        $row['updated']);
             }
         } catch (Exception $e) {
             $ret[] = new User();
@@ -250,10 +250,10 @@ class Article {
         try {
             $tables = array('users_articles' => TBPREFIX.'users_articles');
             $rs = DB::getInstance()->execute(
-                self::DELETE_USER,
-                array(),
-                array($this->id, $idUser),
-                $tables);
+                    self::DELETE_USER,
+                    array(),
+                    array($this->id, $idUser),
+                    $tables);
         } catch (Exception $e) {
             echo 'Caught exception: ',  $e->getMessage(), "\n";
         }
@@ -263,10 +263,10 @@ class Article {
         try {
             $tables = array('users_articles' => TBPREFIX.'users_articles');
             $rs = DB::getInstance()->execute(
-                self::LINK_USER,
-                array(),
-                array($this->id, $idUser),
-                $tables);
+                    self::LINK_USER,
+                    array(),
+                    array($this->id, $idUser),
+                    $tables);
         } catch (Exception $e) {
             echo 'Caught exception: ',  $e->getMessage(), "\n";
         }
@@ -286,10 +286,10 @@ class Article {
         $tables = array("articles" => TBPREFIX."articles");
         try {
             DB::getInstance()->execute(
-                self::INSERT_SQL,
-                array($this->title, $this->subtitle, $this->summary, $this->body, $this->tag, $this->metadescription, $this->metakeyword),
-                array($this->id, $this->number_id, $this->category_id, $this->indexnumber, $this->published, $this->commentsallowed),
-                $tables);
+                    self::INSERT_SQL,
+                    array($this->title, $this->subtitle, $this->summary, $this->body, $this->tag, $this->metadescription, $this->metakeyword),
+                    array($this->id, $this->number_id, $this->category_id, $this->indexnumber, $this->published, $this->commentsallowed),
+                    $tables);
         } catch (Exception $e) {
             echo 'Caught exception: ',  $e->getMessage(), "\n";
         }
@@ -299,10 +299,10 @@ class Article {
         $tables = array("articles" => TBPREFIX."articles");
         try {
             DB::getInstance()->execute(
-                self::UPDATE_SQL,
-                array($this->title, $this->subtitle, $this->summary, $this->body, $this->tag, $this->metadescription, $this->metakeyword),
-                array($this->number_id, $this->category_id, $this->indexnumber, $this->published, $this->commentsallowed, $this->id),
-                $tables);
+                    self::UPDATE_SQL,
+                    array($this->title, $this->subtitle, $this->summary, $this->body, $this->tag, $this->metadescription, $this->metakeyword),
+                    array($this->number_id, $this->category_id, $this->indexnumber, $this->published, $this->commentsallowed, $this->id),
+                    $tables);
         } catch (Exception $e) {
             echo 'Caught exception: ',  $e->getMessage(), "\n";
         }
@@ -336,10 +336,10 @@ class Article {
         $tables = array("articles" => TBPREFIX."articles");
         try {
             $rs = DB::getInstance()->execute(
-                self::SELECT_BY_ID_ORD,
-                array(),
-                array(),
-                $tables);
+                    self::SELECT_BY_ID_ORD,
+                    array(),
+                    array(),
+                    $tables);
             $row = mysql_fetch_array($rs);
             $maxId = $row['id'];
         } catch (Exception $e) {
@@ -353,10 +353,10 @@ class Article {
         $tables = array("articles" => TBPREFIX."articles");
         try {
             $rs = DB::getInstance()->execute(
-                self::SELECT_MAX_INDEXNUMBER,
-                array(),
-                array(),
-                $tables);
+                    self::SELECT_MAX_INDEXNUMBER,
+                    array(),
+                    array(),
+                    $tables);
             if ($row = mysql_fetch_array($rs)) {
                 $maxIndexNumber = $row[0];
             } else {
